requirements.txt
aiofiles
aiohttp
aiohttp_session
cryptography
pyrofork
python-dotenv
tgcrypto